Add convenience method to list color maps.
I find myself often needing to list color maps (mainly for IPython
widgets), and it is painful to get a list of all names[1].
So I think having a list of known colormaps name would be useful

Alternatively this could be a nested dict of category/colorsmaps, but
not sure that is that much useful.

[1]: Sure in widget you can also pass a dict for a dropdown, still I
still wish to be able to filter reversed ones, and list of names.